# apps/api/openapi/school.json
{
  "components": {
    "schemas": {
      "AcademicYearDto": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "endsOn": {
            "format": "date",
            "type": "string"
          },
          "id": {
            "pattern": "^[1-9][0-9]{0,18}$",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "startsOn": {
            "format": "date",
            "type": "string"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AcademicYearStatus"
              }
            ]
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "startsOn",
          "endsOn",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
      "AcademicYearSort": {
        "enum": [
          "-startsOn",
          "startsOn",
          "name",
          "-name"
        ],
        "type": "string"
      },
      "AcademicYearStatus": {
        "enum": [
          "planned",
          "active",
          "closed"
        ],
        "type": "string"
      },
...
      "ApiErrorBodyDto": {
        "properties": {
          "code": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              }
            ]
          },
          "details": {
            "additionalProperties": true,
            "nullable": true,
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "details",
          "requestId"
        ],
        "type": "object"
      },
      "ApiErrorDto": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBodyDto"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "CreateAcademicYearDto": {
        "properties": {
          "endsOn": {
            "description": "After startsOn, and at most 731 days after it",
            "format": "date",
            "type": "string"
          },
          "name": {
            "description": "e.g. 2026-27 or Sept 2026",
            "maxLength": 50,
            "minLength": 2,
            "type": "string"
          },
          "startsOn": {
            "format": "date",
            "type": "string"
          }
        },
        "required": [
          "name",
          "startsOn",
          "endsOn"
        ],
        "type": "object"
      },
...
      "ErrorCode": {
        "enum": [
          "MALFORMED_REQUEST",
          "AUTH_REQUIRED",
          "AUTH_FAILED",
          "PERMISSION_DENIED",
          "SCHOOL_SUSPENDED",
          "ORIGIN_REJECTED",
          "NOT_FOUND",
          "VALIDATION_FAILED",
          "REFERENCE_NOT_FOUND",
          "UNKNOWN_FIELD",
          "INVALID_VALUE",
          "PAYLOAD_TOO_LARGE",
          "UNSUPPORTED_MEDIA_TYPE",
          "RATE_LIMITED",
          "SERVICE_UNAVAILABLE",
          "INTERNAL_ERROR",
          "PASSWORD_CHANGE_REQUIRED",
          "TOTP_REQUIRED",
          "TOTP_NOT_ENROLLED",
          "TOTP_ALREADY_ENROLLED",
          "TOTP_INVALID",
          "CURRENT_PASSWORD_INCORRECT",
          "SCHOOL_SHORT_CODE_TAKEN",
          "SCHOOL_SHORT_CODE_IMMUTABLE",
          "SCHOOL_TERMINATED",
          "ILLEGAL_STATUS_TRANSITION",
          "CONCURRENT_UPDATE",
          "TOKEN_INVALID",
          "EMAIL_NOT_VERIFIED",
          "SELF_ACTION_FORBIDDEN",
          "LAST_PRINCIPAL",
          "IDENTITY_NUMBER_MISSING",
          "ACTIVE_PRINCIPAL_EXISTS",
          "ALREADY_PRINCIPAL",
          "STAFF_NOT_ACTIVE",
          "USER_DISABLED",
          "ACADEMIC_YEAR_NAME_TAKEN",
          "ACADEMIC_YEAR_CLOSED",
          "ACADEMIC_YEAR_HAS_ACTIVE_ENROLMENTS",
          "CLASS_NAME_TAKEN",
          "CLASS_YEAR_IMMUTABLE",
          "CLASS_ARCHIVED",
          "CLASS_HAS_ACTIVE_ENROLMENTS",
          "SECTION_NAME_TAKEN",
          "SECTION_ARCHIVED",
          "SECTION_IN_USE",
          "SUBJECT_NAME_TAKEN",
          "SUBJECT_CODE_TAKEN",
          "SUBJECT_ARCHIVED",
          "GUARDIAN_CNIC_EXISTS",
          "GUARDIAN_CNIC_LOCKED",
          "GUARDIAN_MERGED",
          "GUARDIAN_CNIC_MISSING",
          "GUARDIAN_NO_LOGIN_LINK",
          "GUARDIAN_IS_PRIMARY_CONTACT",
          "LOGIN_ALREADY_EXISTS",
          "LINK_EXISTING_LOGIN_UNCONFIRMED",
          "UNEXPECTED_RESPONSE"
        ],
        "type": "string"
      },
...
      "UpdateAcademicYearDto": {
        "properties": {
          "endsOn": {
            "format": "date",
            "type": "string"
          },
          "name": {
            "maxLength": 50,
            "minLength": 2,
            "type": "string"
          },
          "startsOn": {
            "format": "date",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {},
    "description": "",
    "title": "ASMS school API",
    "version": "1"
  },
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/academic-years": {
      "get": {
        "operationId": "AcademicYearsController_list",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 999999,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 50,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/AcademicYearStatus"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/AcademicYearSort"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "items": {
                        "$ref": "#/components/schemas/AcademicYearDto"
                      },
                      "type": "array"
                    },
                    "limit": {
                      "maximum": 50,
                      "minimum": 1,
                      "type": "integer"
                    },
                    "page": {
                      "minimum": 1,
                      "type": "integer"
                    },
                    "total": {
                      "minimum": 0,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "data",
                    "page",
                    "limit",
                    "total"
                  ],
                  "type": "object"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "academic years"
        ]
      },
      "post": {
        "operationId": "AcademicYearsController_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAcademicYearDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicYearDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "academic years"
        ]
      }
    },
    "/api/v1/academic-years/{id}": {
      "get": {
        "operationId": "AcademicYearsController_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "pattern": "^[1-9][0-9]{0,18}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicYearDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "academic years"
        ]
      },
      "patch": {
        "operationId": "AcademicYearsController_update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "pattern": "^[1-9][0-9]{0,18}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAcademicYearDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicYearDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "academic years"
        ]
      }
    },
    "/api/v1/academic-years/{id}/activate": {
      "post": {
        "operationId": "AcademicYearsController_activate",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "pattern": "^[1-9][0-9]{0,18}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicYearDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "academic years"
        ]
      }
    },
    "/api/v1/academic-years/{id}/close": {
      "post": {
        "operationId": "AcademicYearsController_close",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "pattern": "^[1-9][0-9]{0,18}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicYearDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "academic years"
        ]
      }
    },
...
  },
  "servers": [],
...
}